generations/metroplex-ideaforge-132/requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-click>=1.1.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.1.0
